requirements.txt
ipython>=8.12.3
matplotlib>=3.10.3
matplotlib_inline>=0.1.7
pandas>=2.2.3
Pillow>=11.2.1
scikit_learn>=1.6.1
torch>=2.6.0+cu126
torchvision>=0.21.0+cu126
